fix: move bulk edit tests to BulkEditTest.ts, fix duplicate comment
Move updateMultipleMoneyRequests and bulk edit draft transaction test
blocks from IOUTest.ts to tests/actions/IOUTest/BulkEditTest.ts.
Remove duplicate JSDoc comment on highlightTransactionOnSearchRouteIfNeeded.
